# docs/openapi.yml
openapi: "3.0.3"
info:
  title: Bienenhaus Propiedades API
  description: |
    API RESTful del panel de administración de Bienenhaus Propiedades.
    Inmobiliaria boutique con CRM, tasaciones ACM, portales y redes sociales.
  version: "1.0.0"
  contact:
    name: Bienenhaus Propiedades
    url: https://bienenhaus.com.ar
servers:
  - url: https://bienenhaus.com.ar/api
    description: Producción
  - url: http://localhost:5000/api
    description: Desarrollo
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session
...
paths:
...
  /settings:
    get:
      summary: Obtiene configuración del sitio
      security: [{ SessionAuth: [] }]
      responses:
        "200":
          description: Configuración
    put:
      summary: Actualiza configuración
      security: [{ SessionAuth: [] }]
      responses:
        "200":
          description: Actualizada
...